Create remote file download
This code first checks if the user has submitted a URL through a POST request. If a URL has been submitted, it checks if the URL is valid using the FILTER_VALIDATE_URL filter. If the URL is valid, it uses the basename() function to extract the file name from the URL, and opens a new file for writing.

It then uses the cURL library to download the file from the URL and save it to the local file. Once the download is complete, it closes the file and displays a success message. source belongs to https://androidwell.com/.